About Page, Location Hardcodes, Header, Footer
Added about page. Removed discord logo svg because it wasn't working in vercel and I'll need to look at that more later.

Did a bunch of hardcoding for location page. Will replace with a database pull later.

Styled Header to be fixed to top and spaced out nav items.

Played with the footer, not quite successful yet.